requirements.txt
python-libtorrent>=2.0.10

